requirements.txt
avalanche-lib==0.4.0
neptune==1.8.3
